Cleaned up!
New method for the ++odomoter. Alot easier to understand! using modulo!

Other changes aswell. overall a bit cleaner and nicer.

+++ Chapter_14_more_about_classes/ch14_p15_car_instrument_simulator/fuel_gauge.h
@@ -9,7 +9,7 @@ public:
   Fuel_Gauge();
   Fuel_Gauge operator++();
   Fuel_Gauge operator--();
-  float fuel_report() const { return FuelAmt; }
+  int fuel_report() const { return FuelAmt; }
 };
 
 #endif // FUEL_GAUGE_H
