+++ Chapter_14_more_about_classes/ch14_p15_car_instrument_simulator/odometer.h
@@ -5,12 +5,12 @@
 class Odometer
 {
 private:
-  float CurMileage = 0.0; // max is 999,999
+  int CurMileage = 0; // max is 999,999
   Fuel_Gauge* Tank;
 
 public:
   Odometer(Fuel_Gauge&);
-  float getCurMileage() const { return CurMileage; }
+  int getCurMileage() const { return CurMileage; }
   Odometer operator++();
   void test() { std::cout << Tank->fuel_report(); }
 };
